refactor forms from classes > forms.py
