+++ app/routes.py
@@ -1,6 +1,11 @@
-from io import StringIO
+from datetime import date, datetime
+import io
+import json
 import os
-from datetime import date, datetime, timedelta
+
+from flask import render_template, redirect, url_for, request, jsonify
+from flask_login import current_user, login_user, login_required, logout_user
+import numpy as np
 
 from app import application, db
 from app.utils import *
@@ -8,19 +13,18 @@ from app.classes import (
     User,
     Patient,
     Prescription,
-    PrescriptionForm,
     Patient,
-    PatientForm,
-    RegistrationForm,
-    LogInForm,
-    UploadFileForm,
     Intake,
 )
 
-from flask import render_template, redirect, url_for, request, jsonify
-from flask_login import current_user, login_user, login_required, logout_user
-import numpy as np
-import json
+from app.forms import (
+    PatientForm,
+    RegistrationForm,
+    LoginForm,
+    PrescriptionForm,
+    UploadFileForm,
+)
+
 
 from modeling.preprocessing import *
 
@@ -199,9 +203,7 @@ def add_patient():
 # New user registration -----------------------------------------------------
 @application.route("/register", methods=("GET", "POST"))
 def register():
-    """
-    Registers a user from form data.
-    """
+    """Registers a new doctor (user) from form data."""
     registration_form = RegistrationForm()
     if registration_form.validate_on_submit():
         firstname = registration_form.firstname.data
@@ -214,6 +216,7 @@ def register():
             User.query.filter_by(username=username).count()
             + User.query.filter_by(email=email).count()
         )
+
         if user_count > 0:
             return "<h1>Error - Existing user : " + username + " OR " + email + "</h1>"
         else:
@@ -227,10 +230,8 @@ def register():
 # User login ------------------------------------------------------------------
 @application.route("/login", methods=["GET", "POST"])
 def login():
-    """
-    Logs in a user.
-    """
-    login_form = LogInForm()
+    """Logs in a retturning doctor (user)."""
+    login_form = LoginForm()
     if login_form.validate_on_submit():
         username = login_form.username.data
         password = login_form.password.data
@@ -281,7 +282,7 @@ def adherence_model(
     """
     takes in accelerometer and gyroscope data as a string.
     """
-    raw_sensor_data = StringIO(data)
+    raw_sensor_data = io.StringIO(data)
 
     # run the pill classifier process
     X = preprocess(raw_sensor_data, regression=False)
